requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
openai>=1.6.1,<2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
drf-spectacular==0.27.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
Pillow==10.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0
textract==1.6.5
python-multipart==0.0.6
django-filter==23.3
django-extensions==3.2.3
requests>=2.25.0
google-generativeai>=0.3.0 